vente_pc/requirements.txt
django-resized
uuid
